[REFACTOR] Move files to new locations
